Cleaned up GEC code
